% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_empirical.R
\name{pt_empirical}
\alias{pt_empirical}
\title{Calculating empirical demand indicators}
\usage{
pt_empirical(pt, id_var, max_bp = NULL, pmax = FALSE)
}
\arguments{
\item{pt}{A data frame consisting of the `id_var` and purchase task variables.}

\item{id_var}{The name of the unique identifier as identified in the data frame.}

\item{max_bp}{The value of the assigned breakpoint in the absence of zero consumption by the final price point. The default is calculated empirically
from the data directly, using the formula Pn + (Pn - Pn-1), where Pn is the final price point, and Pn-1 is the price point immediately before the
final price point.}

\item{pmax}{A logical argument whether Pmax should be calculated. The default is FALSE.}
}
\value{
A data frame
}
\description{
This function extracts the empirical demand indicators (Intensity, Breakpoint, and Omax) from the purchase task. The demand indicator Pmax is not calculated
by default, as it is increasingly considered of dubious conceptual interest and to be epiphenomenonological in relation to health outcomes (Martínez-Loredo et al. 2021).
}
\details{
Martínez‐Loredo, V., González‐Roz, A., Secades‐Villa, R., Fernández‐Hermida, J. R., & MacKillop, J. (2021). Concurrent validity of the Alcohol Purchase Task for measuring
the reinforcing efficacy of alcohol: An updated systematic review and meta‐analysis. Addiction, 116(10), 2635-2650.
}
\examples{
### --- Load Data
data("cpt_data")

### --- Prep Data
pt <- price_prep(cpt_data, id_var = "ID", vars = c(paste0("cpt",1:15)),
prices = c("0","0.05","0.10","0.20","0.30","0.40","0.50", "0.75","1","2","3","4","5","7.5","10"))

pt2 <- pt_prep(pt, id_var = "ID", type = "partial", remove0 = TRUE, max_val = 99)
pt3 <- pt_qc(pt2, id_var = "ID", type = "partial")

### --- Function Example
pt4 <- pt_empirical(pt3$data, id_var = "ID")

}
